Reworking country borders - Old borders removed, new ones added
New ones include:
grey areas, solid-, dashed-, dotted borders, major lakes, added in the order mentioned.

The shapefiles for the new borders are loaded starting at line 122. Notes regarding this are in the script after loading the files.

The borders are added to each plot before the "labs" call. For (sub)regional maps, the lakes are added and the grey areas can be as well if the corresponding line is "uncommented".